+++ Include/Entity.hpp
@@ -2,7 +2,7 @@
 #include <string>
 #include <map>
 
-typedef std::map<std::string, size_t> t_map;
+typedef std::map<std::string, double> t_map;
 
 class Entity {
 
@@ -12,14 +12,14 @@ public:
     Entity() = default;
 
 
-    void remaining_HP(size_t damage);
+    void remaining_HP(double damage);
     void stats_change(std::string stat, int change);
     std::string get_status();
 
     t_map get_stats();
     std::string get_name();
-    size_t get_health();
-    size_t get_stat(std::string stat);
+    double get_health();
+    double get_stat(std::string stat);
     void set_status(std::string str);
 
 protected:
